cull imports, fix sylmap center stroke alignment, merge pour and spill, fix center stroke width
