+++ package.json
@@ -2,6 +2,12 @@
   "name": "urbit.avatars",
   "version": "0.1.0",
   "private": true,
+  "devDependencies":{
+    "figma-js": "^1.5.1",
+    "urbit-ob": "^1.0.0",
+    "omit-empty": "^0.4.1",
+    "js-file-download": "^0.4.1"
+  },
   "dependencies": {
     "autoprefixer": "7.1.6",
     "babel-core": "6.26.0",
@@ -12,7 +18,6 @@
     "babel-runtime": "6.26.0",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
-    "chroma-js": "^1.3.6",
     "css-loader": "0.28.7",
     "dotenv": "4.0.0",
     "dotenv-expand": "4.2.0",
@@ -24,25 +29,16 @@
     "eslint-plugin-jsx-a11y": "5.1.1",
     "eslint-plugin-react": "7.4.0",
     "extract-text-webpack-plugin": "3.0.2",
-    "figma-js": "^1.3.4",
     "file-loader": "1.1.5",
-    "firebase": "^5.2.0",
     "fs-extra": "3.0.1",
-    "graphology": "^0.11.4",
-    "graphology-components": "^0.1.0",
-    "hashids": "^1.1.4",
     "html-webpack-plugin": "2.29.0",
     "jest": "20.0.4",
-    "js-combinatorics": "^0.5.3",
-    "js-file-download": "^0.4.1",
     "lodash": "^4.17.10",
     "object-assign": "4.1.1",
-    "omit-empty": "^0.4.1",
     "postcss-flexbugs-fixes": "3.2.0",
     "postcss-loader": "2.0.8",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "re-base": "^3.2.3",
     "react": "^16.3.1",
     "react-dev-utils": "^5.0.1",
     "react-dom": "^16.3.1",
@@ -50,7 +46,6 @@
     "style-loader": "0.19.0",
     "sw-precache-webpack-plugin": "0.11.4",
     "transformation-matrix": "^1.11.0",
-    "urbit-ob": "^1.0.0",
     "url-loader": "0.6.2",
     "webpack": "3.8.1",
     "webpack-dev-server": "2.9.4",
