+++ src/views/Pour.js
@@ -1,19 +1,23 @@
 import React, { Component } from 'react'
 import { toPlanetName } from 'urbit-ob'
+import { map } from 'lodash'
 
 import {
   randomShip,
 } from '../lib/lib'
 
+import {
+  seq,
+} from '../lib/lib.array'
+
 import { pour } from '../lib/pour'
 
-import { ReactSVGComponents } from '../renderers/ReactSVGComponents'
+import ReactSVGComponents from '../renderers/ReactSVGComponents'
 
 import sylmap from '../sylmaps/sylmap'
 
 import { InputBox } from '../components/UI'
 
-const SIZE = 600
 
 class Pour extends Component {
   constructor(props) {
@@ -39,6 +43,7 @@ class Pour extends Component {
     const { patp } = this.state
     return (
       <div>
+
         <button onClick={() => this.setState({ patp: toPlanetName(randomShip('planet')) })}>
           {'Random @P'}
         </button>
@@ -46,19 +51,40 @@ class Pour extends Component {
           placeholder={'~ridlur-figbud'}
           title={'Submit'}
           submit={(content) => {
-            console.log(content)
             this.setState({patp: content})
           }}
         />
         {
           pour({
-            patp: this.state.patp,
+            patp: patp,
             sylmap: sylmap,
             renderer: ReactSVGComponents,
             size: 256,
           })
         }
-        <p>{this.state.patp}</p>
+          <p>{patp}</p>
+
+
+        <div className={'flex-c w100'}>
+        <div className={'breakAfter'}>
+          {
+            map(seq((3 * 4 * 16)), n => {
+              const randPlanet = toPlanetName(randomShip('planet'))
+              return <div className={'ml-2 pba nob'}>
+              {pour({
+                patp: randPlanet,
+                sylmap: sylmap,
+                renderer: ReactSVGComponents,
+                size: 256,
+                colorway: ['white', 'black'],
+              })}
+              <p className={'mono tssub'}>{randPlanet}</p>
+              </div>
+            })
+          }
+
+        </div>
+        </div>
       </div>
     )
   }
